# backend/src/main/resources/openapi/mobile-v1.yaml
openapi: 3.0.3
info:
  title: AI Workdeck Mobile Relay API
  version: "1.0.0"
  description: |
    手机端（iOS / 小程序 / 安卓 / 鸿蒙）与桌面端共用的中转接口。真源在本文件；
    移动仓 contract/api/mobile-v1.yaml 是钉版副本。鉴权一律请求头 X-Session-Id。
    未登录不是 401，而是 HTTP 200 + {code: 4010} 信封——客户端按信封判鉴权失败。

    校验器（swagger-request-validator）默认 additionalProperties: false：本文件是响应字段的
    **穷举白名单**，服务端新增任何响应字段都必须先写进这里，否则契约测试红。
servers:
  - url: https://addin.aiworkdeck.com
    description: 大陆版（北京）
  - url: https://addin.workdeck.ai
    description: 国际版（新加坡）
components:
  parameters:
    SessionId:
      name: X-Session-Id
      in: header
      required: false
      schema: { type: string }
      description: 登录会话或 awdt_ 设备令牌。缺失或失效 → 200 + Envelope(code 4010)。
  schemas:
    Envelope:
      type: object
      required: [code]
      properties:
        code: { type: integer, description: "0 成功；4010 未登录；1 业务错误" }
        message: { type: string, nullable: true }
        data: { nullable: true }
        feature: { type: string, description: "GlobalExceptionHandler code=4001/4003 附带的功能 id" }
        configured: { type: boolean, description: "code=4001（功能未配置）附带" }
        usage: { type: object, additionalProperties: true, description: "code=4003（额度已满）附带的用量明细" }
        gatewayKind: { type: string, description: "code=1（平台网关失败）附带" }
        canUseOwnKey: { type: boolean, description: "code=1（平台网关失败）附带，是否可用自备 Key 绕过" }
        kind:
          type: string
          enum: [DISABLED, UNAVAILABLE, NOT_CONNECTED, NOT_FOUND, REJECTED, REVIEW_ACCOUNT, ALREADY_PAID, IDEMPOTENCY_CONFLICT]
          description: |
            code=1 时 /api/mobile/billing/* 附带的**机器可读判别位**（dev-board#425）。
            **四端一律按它分支，禁止匹配 message 措辞**——message 经服务端 LangText 在英文部署下
            会整条变成英文，按中文串做分支必然落空；code 也永远是 1，判 code 分不出这八种情况。
            message 只用于兜底展示。

            每条末尾方括号里是**余额那一行**的渲染裁定，唯一来源是移动仓
            contract/schema/billing.schema.json 的说明段（四端逐条对齐，各自有测试钉住）。
            判据是**这一态会不会自己恢复**：终态一律整行不渲染，绝不能渲染成
            「余额暂时读不到，稍后再试」——那是让用户去重试一句永不改变的报错。
              DISABLED             本服务器没配 mobile.billing.*，压根没开通
                                   [整行不渲染，充值入口一并收起]。**未配 env 即此态，也就是四端一发版时的默认生产状态**
              UNAVAILABLE          上游不可用（网络/5xx/官网空体 404 即密钥或 env 配错）
                                   [显示 balance.unavailable，稍后重试]，**绝不能显示成「没有账户」或「余额 0」**
              NOT_CONNECTED        这个登录账号还没有统一账户（无已验证身份，或官网按该身份查无账户）
                                   [整行不渲染]。读余额一律 create=false，服务端不会替用户建号
              NOT_FOUND            官网明确回「查无此物」（accountId 已注销 / 单号不属于该账户）
                                   [显示 balance.unavailable]
              REJECTED             官网明确拒绝的业务错误，或统一账户已绑给别的登录账号
                                   [显示 balance.unavailable]
              REVIEW_ACCOUNT       App 审核演示账号，余额与充值一律不可用
                                   [整行不渲染]。审核员进设置页看到的必须是干净的一页，不是一行报错
              ALREADY_PAID         该幂等键的单子已支付 → 带 outTradeNo 转去查单，别重新下单
                                   [与余额无关：只出现在下单/查单路径]
              IDEMPOTENCY_CONFLICT 同一把幂等键换了金额/类型 → 带 outTradeNo [同上]
              （kind 缺席）        非 billing 专有的失败走通用 handler
                                   [显示 balance.unavailable]，不得当成「没有账户」
        outTradeNo:
          type: string
          description: |
            仅 kind=ALREADY_PAID / IDEMPOTENCY_CONFLICT 时出现：官网连同 409 一起回的既有商户订单号。
            这是「App 被杀、本地没存下单号」的恢复路径，客户端拿它调 /recharge/status 查单。
    RelayProject:
      type: object
      required: [deviceId, key, name]
      properties:
        deviceId: { type: string, description: 桌面机 UUID }
        deviceName: { type: string, nullable: true }
        key: { type: string, description: 桌面机本地库项目 id；跨机同号不同物，必须连 deviceId }
        name: { type: string }
    MediaStatus:
      type: object
      required: [clientMediaId, delivered, waitingSeconds]
      properties:
        clientMediaId: { type: string }
        delivered: { type: boolean, description: 桌面端已 ACK 落盘（中转区已删） }
        waitingSeconds: { type: integer, format: int64, description: 已投递为 0 }
        expiresAt: { type: string, description: 中转区到期时刻（ISO 本地时间），仅未投递件带 }
    MediaUsage:
      type: object
      required: [usedBytes, quotaBytes]
      description: |
        usedBytes 只统计 inbox 待取件字节（sumPendingBytes）；而配额执行（storeMedia 时的 3GB 判断）
        同时计入 inbox 与 transfer。手机端设置页显示的用量可能低于实际被计费的量。服务端行为，本轮不改。
      properties:
        usedBytes: { type: integer, format: int64 }
        quotaBytes: { type: integer, format: int64 }
    UploadResult:
      type: object
      required: [code, id, clientMediaId, delivered]
      properties:
        code: { type: integer, enum: [0] }
        id: { type: integer, format: int64 }
        clientMediaId: { type: string }
        delivered: { type: boolean, description: 幂等重传命中已投递记录时为 true }
    AccountUser:
      type: object
      required: [id, username, displayName]
      properties:
        id: { type: integer, format: int64 }
        username: { type: string }
        displayName: { type: string }
        avatarUrl: { type: string, description: 无头像时为空串 }
        role: { type: string }
        subscriptionType: { type: string, nullable: true }
    LoginData:
      type: object
      required: [sessionId, isNewUser, user]
      description: |
        sms-login/verify 与 wx-phone-login 成功时的 data（两条路逐字同形）。
        isNewUser 由服务端恒返回（account.created）。
      properties:
        sessionId: { type: string }
        isNewUser: { type: boolean }
        mustBindPhone: { type: boolean, description: 由 account-login / mail-login 等门控登录流设置；sms-login/verify 与 wx-phone-login 不返回 }
        user: { $ref: '#/components/schemas/AccountUser' }
    LoginEnvelope:
      type: object
      required: [code]
      properties:
        code: { type: integer }
        message: { type: string, nullable: true }
        data: { $ref: '#/components/schemas/LoginData' }
    BillingBalance:
      type: object
      description: |
        统一账户余额（dev-board#425）。权威在官网仓（credit_lots + wallet_ledger），
        云后端经内部记账口 POST /api/internal/account 读取，服务端带 30 秒 TTL 缓存。

        **展示口径（四端逐字符一致）**：符号 + balanceCents/100 固定两位小数、**不带千分位**、
        **不跟随设备 locale**；符号按 currency 取（CNY→¥、USD→$），任何一端都不得写死 ¥。
        唯一来源是移动仓 contract/schema/billing.schema.json，期望串在 fixtures/billing.json
        的 balance[].display（三端各自的格式化函数与它对拍）。
      required: [balanceCents, currency, plan]
      properties:
        balanceCents: { type: integer, format: int64, description: 余额，整数分（Credits 1 元 = 100 分） }
        currency: { type: string, enum: [CNY, USD], description: 站点币种，由部署站点决定 }
        plan:
          type: string
          nullable: true
          description: |
            计费档位，权威侧 lib/wallet.ts 的 getPlan() 只返回 `paid` / `free`——**不是套餐名**，
            也基本不会为 null（只有上游没给这个字段时才是 null）。任何一端都不要把它当套餐名直接渲染。
    RechargeOrder:
      type: object
      description: |
        充值单。present=qrcode 时 codeUrl/qrCode 有值，present=redirect 时 redirectUrl 有值，
        present=virtual 时 signData/paySig/signature 三者都有值，present=native 时 appAccountToken 有值；
        不适用的字段不出现在响应里。

        present=virtual 是小程序虚拟支付（dev-board#427）：三个字段由官网算出后经云后端原样透传，
        小程序拿它们调 wx.requestVirtualPayment({signData, paySig, signature, mode:'short_series_goods'})。
        云后端不持有虚拟支付 AppKey、不碰 session_key，全程只当管道。

        present=native 是 iOS App Store 内购（dev-board#426）：客户端拿 appAccountToken 调
        product.purchase(options:[.appAccountToken(UUID)])，成功后把 Transaction.jwsRepresentation
        交给 POST /api/mobile/billing/recharge/confirm；**收到 paid 之前绝不 tx.finish()**。
        验签在官网（苹果根证书 / bundleId / appAppleId 都在那边），云后端同样只当管道。
      required: [present, outTradeNo, amountCents]
      properties:
        present: { type: string, enum: [qrcode, redirect, virtual, native] }
        outTradeNo: { type: string, description: 商户订单号，查状态用 }
        amountCents: { type: integer, format: int64, description: 下单金额，整数分 }
        codeUrl: { type: string, description: present=qrcode：支付链接原文 }
        qrCode: { type: string, description: present=qrcode：二维码图（data URI） }
        redirectUrl: { type: string, description: present=redirect：跳转地址 }
        signData: { type: string, description: 'present=virtual 必有：待签名 JSON 串（offerId/buyQuantity/env/currencyType/productId/goodsPrice/outTradeNo/attach）' }
        paySig: { type: string, description: 'present=virtual 必有：hex(hmac_sha256(appKey, "requestVirtualPayment&" + signData))' }
        signature: { type: string, description: 'present=virtual 必有：hex(hmac_sha256(session_key, signData))' }
        appAccountToken:
          type: string
          description: |
            present=native 必有：官网为这张单生成的 UUID（订单的 providerRef）。iOS 把它塞进
            purchase 的 appAccountToken 选项，苹果在 signedTransaction 里原样带回，
            官网据此把交易认回这张单（App 被杀、本地没存下单号时的唯一线索）。
    RechargeStatus:
      type: object
      required: [status, paid, amountCents]
      properties:
        status: { type: string, enum: [pending, paid, closed, expired] }
        paid: { type: boolean, description: 已到账；置 true 后服务端作废该用户的余额缓存 }
        amountCents: { type: integer, format: int64 }
    CaptureManifest:
      type: object
      description: |
        客户端随影像生成的取证清单（与 iOS CaptureManifest.swift 逐字段一致）。当前服务端只接收
        clientMediaId / capturedAt / fileName / mediaType 四个字段（multipart 参数），其余留在客户端；
        接 TSA 后 tsaToken 上行，结构不变。
      required: [clientMediaId, sha256, capturedAt, deviceModel, osVersion, appVersion, fromCamera]
      properties:
        clientMediaId: { type: string, format: uuid, description: 幂等键 }
        sha256: { type: string, minLength: 64, maxLength: 64 }
        capturedAt: { type: string, format: date-time, description: 设备时钟 }
        serverReceivedAt: { type: string, format: date-time, nullable: true }
        latitude: { type: number, nullable: true }
        longitude: { type: number, nullable: true }
        horizontalAccuracy: { type: number, nullable: true, description: 米 }
        deviceModel: { type: string }
        osVersion: { type: string }
        appVersion: { type: string }
        fromCamera: { type: boolean, description: 是否强制来自相机 }
        tsaToken: { type: string, nullable: true, description: 可信时间戳，当前恒 null }
paths:
  /api/auth/sms-login/send-code:
    post:
      summary: 发送短信验证码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties: { phone: { type: string, pattern: '^1\d{10}$' } }
      responses:
        "200":
          description: code 0 已发送；code 1 业务拒绝（频率、号码格式、人机门）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
  /api/auth/sms-login/verify:
    post:
      summary: 校验验证码并签发会话
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, code]
              properties:
                phone: { type: string }
                code: { type: string }
      responses:
        "200":
          description: code 0 带 data；否则 code 非 0 且 data 缺失
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginEnvelope' }
  /api/auth/wx-phone-login:
    post:
      summary: 微信手机号一键登录（匿名）
      description: |
        小程序 button open-type="getPhoneNumber" 回调里的一次性 code 换手机号并签发会话
        （dev-board#534）。换号在官网侧完成（内部记账口 action=wx-phone），云后端不持有
        小程序 AppSecret。成功响应与 sms-login/verify 逐字同形（LoginData，含 isNewUser）。

        失败一律 200 + code 1 + message，**不带 kind**：客户端 toast 该 message 并展开
        短信登录表单即可，不需要按机器可读位分支（与 /api/mobile/billing/* 不同）。
        三种可预期的失败——本服务器/官网未开通微信登录、code 已过期、非中国大陆手机号。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string, description: getPhoneNumber 回调的 e.detail.code，一次性 }
      responses:
        "200":
          description: code 0 带 data；否则 code 1 且 data 缺失
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginEnvelope' }
  /api/mobile/projects:
    get:
      summary: 手机端：该账号全部设备的项目目录并集（裸数组）
      parameters: [{ $ref: '#/components/parameters/SessionId' }]
      responses:
        "200":
          description: 裸数组；未登录为 Envelope(4010)
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items: { $ref: '#/components/schemas/RelayProject' }
                  - $ref: '#/components/schemas/Envelope'
    put:
      summary: 桌面端：项目目录全量替换（本期不校验）
      parameters: [{ $ref: '#/components/parameters/SessionId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [deviceId, projects]
              properties:
                deviceId: { type: string }
                deviceName: { type: string, nullable: true }
                projects:
                  type: array
                  items:
                    type: object
                    required: [key, name]
                    properties: { key: { type: string }, name: { type: string } }
      responses:
        "200":
          description: code 0 + count/totalCount/truncated
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    required: [code, count, totalCount, truncated]
                    properties:
                      code: { type: integer, enum: [0] }
                      count: { type: integer }
                      totalCount: { type: integer }
                      truncated: { type: boolean }
                  - $ref: '#/components/schemas/Envelope'
  /api/mobile/media:
    post:
      summary: 手机端：上传一件现场影像到中转区（幂等键 clientMediaId）
      parameters: [{ $ref: '#/components/parameters/SessionId' }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, deviceId, projectKey, clientMediaId, fileName, mediaType]
              properties:
                file: { type: string, format: binary }
                deviceId: { type: string }
                projectKey: { type: string }
                clientMediaId: { type: string }
                fileName: { type: string }
                mediaType: { type: string, enum: [image, video, audio] }
                capturedAt: { type: string, description: ISO-8601，可选 }
      responses:
        "200":
          description: 成功 UploadResult；未登录 Envelope(4010)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/UploadResult'
                  - $ref: '#/components/schemas/Envelope'
  /api/mobile/media/status:
    get:
      summary: 手机端：查询影像投递状态（裸数组）
      parameters:
        - { $ref: '#/components/parameters/SessionId' }
        - name: clientMediaIds
          in: query
          required: true
          schema: { type: string }
          description: 逗号分隔
      responses:
        "200":
          description: 裸数组；未登录 Envelope(4010)
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items: { $ref: '#/components/schemas/MediaStatus' }
                  - $ref: '#/components/schemas/Envelope'
  /api/mobile/media/usage:
    get:
      summary: 手机端：中转区用量与配额（裸对象）
      parameters: [{ $ref: '#/components/parameters/SessionId' }]
      responses:
        "200":
          description: 裸对象；未登录 Envelope(4010)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/MediaUsage'
                  - $ref: '#/components/schemas/Envelope'
  /api/mobile/devices:
    get:
      summary: 插件端：设备清单（本期不校验）
      parameters: [{ $ref: '#/components/parameters/SessionId' }]
      responses:
        "200":
          description: 裸数组
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      type: object
                      required: [deviceId, online, projects]
                      properties:
                        deviceId: { type: string }
                        deviceName: { type: string, nullable: true }
                        lastSeenAt: { type: string, nullable: true, description: 最后心跳时刻（ISO 本地时间），无心跳为 null }
                        online: { type: boolean }
                        projects:
                          type: array
                          items:
                            type: object
                            required: [key, name]
                            properties: { key: { type: string }, name: { type: string } }
                  - $ref: '#/components/schemas/Envelope'
  /api/mobile/inbox:
    get:
      summary: 桌面端：本设备待取件（本期不校验）
      parameters:
        - { $ref: '#/components/parameters/SessionId' }
        - { name: deviceId, in: query, required: true, schema: { type: string } }
      responses:
        "200":
          description: 裸数组
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      type: object
                      required: [id, projectKey, clientMediaId, fileName, mediaType, fileSize, createdAt]
                      properties:
                        id: { type: integer, format: int64 }
                        projectKey: { type: string }
                        clientMediaId: { type: string }
                        fileName: { type: string }
                        mediaType: { type: string }
                        fileSize: { type: integer, format: int64 }
                        capturedAt: { type: string, nullable: true }
                        createdAt: { type: string }
                  - $ref: '#/components/schemas/Envelope'
  /api/mobile/inbox/{id}/content:
    get:
      summary: 桌面端：取件字节流。契约红线：成功必须 2xx + application/octet-stream 裸字节，不许 302（本期不校验）
      parameters:
        - { $ref: '#/components/parameters/SessionId' }
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        "200":
          description: 字节流；未登录为 JSON Envelope(4010)
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
  /api/mobile/inbox/{id}/ack:
    post:
      summary: 桌面端：确认落盘，置 deliveredAt 并立即删 blob（本期不校验）
      parameters:
        - { $ref: '#/components/parameters/SessionId' }
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        "200":
          description: code 0
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
  /api/mobile/billing/balance:
    get:
      summary: 手机端：统一账户余额（裸对象）
      parameters: [{ $ref: '#/components/parameters/SessionId' }]
      responses:
        "200":
          description: |
            裸对象；未登录 Envelope(4010)；未绑定/无已验证身份/审核账号/官网不可达
            一律 Envelope(code 1) + kind + 可读 message（绝不用余额 0 冒充）。
            **读余额永不在官网建号**（内部口 resolve 传 create=false）：查无此账户回
            kind=NOT_CONNECTED，客户端据此不渲染余额那一行。建号只发生在用户显式发起充值时。
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/BillingBalance'
                  - $ref: '#/components/schemas/Envelope'
  /api/mobile/billing/recharge:
    post:
      summary: 手机端：创建充值单（裸对象）
      parameters: [{ $ref: '#/components/parameters/SessionId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amountCents, idempotencyKey]
              properties:
                amountCents: { type: integer, format: int64, description: 充值金额，整数分，须 > 0 }
                idempotencyKey:
                  type: string
                  pattern: '^[A-Za-z0-9_-]{8,64}$'
                  description: |
                    **客户端生成并落盘后传入**（扛 App 被杀）。服务端不代生成，缺失即 code 1。
                    官网 orders 表有 UNIQUE(userId, idempotencyKey) 兜底。
                channel:
                  type: string
                  enum: [wxvp, appstore]
                  description: |
                    支付通道（dev-board#427/#426）。**缺省 = 站点默认通道**（第一期行为，扫码/跳转）；
                    wxvp = 微信小程序虚拟支付，**channel=wxvp 时 channel/productId/wxCode 三者必填**，
                    响应为 present=virtual；appstore = iOS 内购，**productId 必填、wxCode 不传**，
                    响应为 present=native。别的取值一律 code 1（通用 handler，无 kind）。
                productId:
                  type: string
                  pattern: '^[a-z0-9_.]{1,64}$'
                  description: |
                    channel=wxvp 时必填：微信道具 id（credits_cny_10 / credits_cny_50 / credits_cny_100）；
                    channel=appstore 时必填：App Store 商品 id，**带点号**
                    （credits.cny.50 / credits.cny.100 / credits.cny.300）。
                    服务端按通道分别校形态：wxvp 是 `^[a-z0-9_]{1,64}$`，appstore 是 `^[a-z0-9_.]{1,64}$`。
                    **价格权威在官网**，云后端只校形态；档位与金额不符时官网回 400 product_mismatch
                    → 信封 kind=REJECTED。message 目前逐字是「充值档位与金额不符，请更新小程序后重试」
                    （两个通道共用这一句，iOS 上措辞不贴切）——**客户端一律按 kind 分支，别读 message**。
                wxCode:
                  type: string
                  description: |
                    channel=wxvp 时必填：wx.login() 拿到的一次性 code，官网用它换 openid + session_key
                    算签名。**session_key 只在官网内存里用，不落库、不打日志**，云后端只透传 code。
      responses:
        "200":
          description: 裸对象；未登录 Envelope(4010)；业务错误 Envelope(code 1)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/RechargeOrder'
                  - $ref: '#/components/schemas/Envelope'
  /api/mobile/billing/recharge/confirm:
    post:
      summary: 手机端：确认 App Store 内购交易（裸对象）
      description: |
        iOS 内购（dev-board#426）：product.purchase 拿到 .success(.verified(tx)) 之后调这里，
        云后端透传官网 confirm-appstore，官网验签（bundleId / appAppleId / 苹果根证书）+ 校
        appAccountToken 与 productId + 判 transactionId 未用过，然后入账。
        **客户端收到 paid 之前绝不 tx.finish()**——finish 掉的交易苹果不再重放。

        受 mobile.billing.recharge-enabled 总开关约束（关时 Envelope kind=DISABLED）。
        失败判别位：验签失败 / 订单不符（官网 400 transaction_invalid）与「该交易已使用」
        （官网 409 transaction_reused）都是 kind=REJECTED，只有 message 不同；
        查无此单（官网 404）是 kind=NOT_FOUND；上游故障是 kind=UNAVAILABLE。
      parameters: [{ $ref: '#/components/parameters/SessionId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [signedTransaction]
              properties:
                outTradeNo:
                  type: string
                  pattern: '^[A-Za-z0-9_-]{1,64}$'
                  description: |
                    **可缺省**：App 被杀后重放 Transaction.unfinished 时本地可能没存下单号，
                    这时只带 signedTransaction 来，官网用交易里的 appAccountToken 反查 providerRef。
                signedTransaction:
                  type: string
                  maxLength: 16384
                  description: |
                    StoreKit 2 的 Transaction.jwsRepresentation 原文。云后端**不验签**、不解析，
                    原样转给官网；超过 16KB 即 code 1。
      responses:
        "200":
          description: 裸对象（形状同 /recharge/status）；未登录 Envelope(4010)；业务错误 Envelope(code 1)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/RechargeStatus'
                  - $ref: '#/components/schemas/Envelope'
  /api/mobile/billing/recharge/status:
    get:
      summary: 手机端：查充值单状态（裸对象）
      parameters:
        - { $ref: '#/components/parameters/SessionId' }
        - { name: outTradeNo, in: query, required: true, schema: { type: string } }
      responses:
        "200":
          description: 裸对象；未登录 Envelope(4010)；业务错误 Envelope(code 1)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/RechargeStatus'
                  - $ref: '#/components/schemas/Envelope'